DEPS - Adding Freenove Dependencies and Installation
Added a submodule which provides the dependencies needed for the hexapod
robot; ensure to `git submodule update --init --recursive` after cloning
this repo.

Also updated the Makefile to install these dependencies in the Arduino
user library directory.  Note these libraries depend on standard Arduino
libraries which have to be explicitly listed in the Makefile; their
installation is handled.